Refs #7614 fix feature specs

+++ db/migrate/20180717124110_associate_api_keys_to_workbench.rb
@@ -1,7 +1,7 @@
 class AssociateApiKeysToWorkbench < ActiveRecord::Migration
   def change
     remove_column :api_keys, :referential_id
-    add_column :api_keys, :workbench_id, :integer
+    add_column :api_keys, :workbench_id, :bigint
 
     ApiKey.where(workbench_id: nil).each do |key|
       key.update_column(:workbench_id, Organisation.find(key[:organisation_id]).workbenches.first&.id)
